+++ index.js
@@ -5,12 +5,13 @@ const http = require('https')
 const csv = require('csv')
 const fs = require('fs')
 const path = require('path')
+const { getopt } = require('stdio')
 
 const KEY_COLUMN_POS = 0
 const DEFAULT_COLUMN_POS = 1 // Column with default language
 
 const opt = {
-  'key': process.env.TRANSLATION_KEY,
+  'key': process.env.GSTRANSLATE_KEY || process.env.TRANSLATION_KEY,
   'pages': (process.env.TRANSLATION_PAGES || '').split(/\s*,\s*/).filter(NotEmpty),
   'pretty-print': parseBool(process.env.TRANSLATION_PRETTY || '') ? 1 : 0,
   'out-dir': path.normalize(
@@ -22,6 +23,68 @@ const opt = {
   'delay': 50
 }
 
+const args = getopt({
+  'key': {
+    key: 'k',
+    args: 1,
+    default: '', // TODO: Remove for version 2 (breaking change)
+    // required: true, // TODO: Add for version 2 (breaking change)
+    description: 'Key of google spreadsheet document'
+  },
+  'pages': {
+    key: 'p',
+    default: [], // TODO: Remove for version 2 (breaking change)
+    // required: true, // TODO: Add for version 2 (breaking change)
+    multiple: true,
+    description: 'Pages list in google spreadsheet document that you wish to generate translations'
+  },
+  'pretty-print': {
+    // default: false, // TODO: Add for version 2 (breaking change)
+    description: 'Pretty-print output translations'
+  },
+  'out-dir': {
+    key: 'o',
+    args: 1,
+    default: 'undefined', // TODO: Change to './locales' for version 2 (breaking change)
+    description: 'Output directory to place generated translation files'
+  },
+  'delay': {
+    args: 1,
+    default: 50,
+    description: 'Delay in ms between requests to pages'
+  }
+})
+// Backward computability for migrate from env (args rewrites env)
+
+if ([args.pages].flat().length > 0) {
+  opt.pages = [args.pages].flat()
+}
+if (args.key) {
+  opt.key = args.key
+}
+opt.delay = Number.parseInt(args.delay)
+if (args['out-dir'] !== 'undefined') {
+  opt["out-dir"] = args['out-dir']
+}
+if (args['pretty-print']) {
+  opt['pretty-print'] = args['pretty-print']
+}
+
+// Deprecations
+if (process.env.TRANSLATION_KEY) {
+  console.warn('\x1b[33m[WARN]\x1b[0m Env \x1b[1mTRANSLATION_KEY\x1b[0m is deprecated, use \x1b[1mGSTRANSLATE_KEY\x1b[0m instead')
+}
+if (process.env.TRANSLATION_PAGES) {
+  console.warn('\x1b[33m[WARN]\x1b[0m Env \x1b[1mTRANSLATION_PAGES\x1b[0m is deprecated, please pass pages list via cli arguments (\x1b[1m--pages|-p\x1b[0m)')
+}
+if (process.env.TRANSLATION_PRETTY) {
+  console.warn('\x1b[33m[WARN]\x1b[0m Env \x1b[1mTRANSLATION_PRETTY\x1b[0m is deprecated, please use cli argument \x1b[1m--pretty-print\x1b[0m instead')
+}
+if (process.env.TRANSLATION_DIR) {
+  console.warn('\x1b[33m[WARN]\x1b[0m Env \x1b[1mTRANSLATION_DIR\x1b[0m is deprecated, please use cli argument \x1b[1m--out-dir\x1b[0m instead')
+}
+
+// TODO: Remove below checks in case get rid of env in version 2
 if (opt.pages.length <= 0) {
   console.error('Pages is not specified. use env variable TRANSLATION_PAGES to specify list of pages separated by comma')
   process.exit(1)